requirements.txt
PyQt5
PyOpenGL
PyOpenGL_accelerate
ezdxf
matplotlib
scipy
tqdm